+++ src/layer.h
@@ -8,17 +8,17 @@
 class layer {
 public:
     layer()  {}
-    layer(int s, Vec2f p) : num_neurons(s), pos(p) {}
+    layer(int s, Vec2d p) : num_neurons(s), pos(p) {}
     int num_neurons;
-    Vec2f pos;
+    Vec2d pos;
     int neuron_size = 30;
     std::vector<neuron> n;
-    float total_height;
+    double total_height;
     bool hovering = false;
-    const float scale = 2.5;
+    const double scale = 2.5;
     void init();
-    void update_pos(Vec2f p);
-    void update_value(std::vector<float> values);
+    void update_pos(Vec2d p);
+    void update_value(std::vector<double> values);
     void draw();
 };
